Split preview into modular components
